Use a better vocabulary
`recognize` does not make explicit we are executing any code.

So it is better to name the function that call the page action of
`dispatch`. This function is responsible to call `recognize` to
recognize the current scope and to dispatch the calls for all the
functions registered to that scope.